Current without JWT

+++ tamu-dues-api/src/main/java/com/venler42/tamu_dues_api/repository/AssignmentRepository.java
@@ -7,10 +7,10 @@ import com.venler42.tamu_dues_api.model.Assignment;
 import java.util.List;
 import java.util.Optional;
 
-public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
-    List<Assignment> findByUserId(Long userId); // key word findBy is critical so JPA knows its query method (in
-                                                // function
-                                                // UserId looks for userId)
+public interface AssignmentRepository extends JpaRepository<Assignment, Integer> {
+    List<Assignment> findByUserId(Integer userId); // key word findBy is critical so JPA knows its query method (in
+                                                   // function
+                                                   // UserId looks for userId)
 
-    Optional<Assignment> findByIdAndUserId(Long assignmentId, Long userId);
+    Optional<Assignment> findByIdAndUserId(Integer assignmentId, Integer userId);
 }
