+++ tamu-dues-api/src/main/java/com/venler42/tamu_dues_api/repository/SectionRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.venler42.tamu_dues_api.model.Section;
 
-public interface SectionRepository extends JpaRepository<Section, Long> {
+public interface SectionRepository extends JpaRepository<Section, Integer> {
 }
